backend/requirements.txt
# Web & API
fastapi>=0.110,<0.114
uvicorn[standard]>=0.23
watchfiles>=0.21

# DB & config
SQLAlchemy>=2.0
python-dotenv>=1.0
PyMySQL>=1.1

# Auth / security
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt>=4.0

# Uploads / HTTP
python-multipart>=0.0.9
httpx>=0.27

# PDF generation & charts
reportlab>=3.6
matplotlib>=3.8
pillow>=10.0

# PDF parsing (used by extractor)
pdfminer.six>=20231228
pypdf>=4.2

# Validation & Schema
pydantic[email]>=2.0
email-validator>=2.2.0

# ---- Optional database drivers (uncomment if needed) ----
# psycopg2-binary>=2.9   # PostgreSQL
# sqlite3                # built-in with Python; no package needed
